requirements.txt
discord.py
python-dotenv
yt-dlp
PyNaCl
aiohttp
async-timeout
